More of Ian's patches -- these are mainly int->ssize_t changes, since those two types are not compatible on the Mac.
git-svn-id: svn://okws.org/trunk/okws2@1326 7d287422-6eea-da11-b677-00123f2a783a

+++ client/mygz2.C
@@ -76,7 +76,7 @@ go (const str &s, mstr *m, u_int out_avail)
 
   u_int p;
   for (p = 0; p <= s.len (); p += chnk) {
-    len = min (chnk, (s.len () - p));
+    len = min (chnk, static_cast<u_int> ((s.len () - p)));
     inb = const_cast<Bytef *> (reinterpret_cast<const Bytef *> 
 			       (s.cstr () + p));
     stream.next_in = inb;
